major cleanup refactor
